+++ package.json
@@ -41,8 +41,6 @@
     "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-config-prettier": "8.6.0",
-    "enzyme": "^3.11.0",
-    "enzyme-to-json": "^3.6.2",
     "@typescript-eslint/eslint-plugin": "^5.49.0",
     "@typescript-eslint/parser": "^5.49.0",
     "prettier": "2.8.3",
@@ -60,7 +58,6 @@
     "jest-canvas-mock": "2.4.0",
     "ts-jest":"29.0.3",
     "serve": "^14.1.2",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
     "whatwg-fetch": "^3.6.2",
     "babel-polyfill": "6.26.0",
     "identity-obj-proxy": "^3.0.0",
